--- app/src/main/java/com/techcourse/ManualHandlerMapping.java
@@ -1,34 +1,0 @@
-package com.techcourse;
-
-import com.techcourse.controller.*;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import jakarta.servlet.http.HttpServletRequest;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
-import webmvc.org.springframework.web.servlet.mvc.asis.ForwardController;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerMapping;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class ManualHandlerMapping implements HandlerMapping {
-
-    private static final Logger log = LoggerFactory.getLogger(ManualHandlerMapping.class);
-
-    private static final Map<String, Controller> controllers = new HashMap<>();
-
-    @Override
-    public void initialize() {
-        log.info("Initialized Handler Mapping!");
-        controllers.keySet()
-                .forEach(path -> log.info("Path : {}, Controller : {}", path, controllers.get(path).getClass()));
-    }
-
-    @Override
-    public Object getHandler(final HttpServletRequest request) {
-        final String requestURI = request.getRequestURI();
-        log.debug("Request Mapping Uri : {}", requestURI);
-        return controllers.get(requestURI);
-    }
-}
