+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/DispatcherServlet.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet.mvc.tobe;
 
 import java.util.List;
 
@@ -7,19 +7,10 @@ import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
-import org.apache.catalina.core.StandardContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.mvc.tobe.AnnotationHandlerMapping;
-import webmvc.org.springframework.web.servlet.mvc.tobe.ControllerAdopter;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerAdopter;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerAdopters;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerExecutionAdopter;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerMapping;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerMappings;
-import webmvc.org.springframework.web.servlet.view.JspView;
 
 public class DispatcherServlet extends HttpServlet {
 
@@ -32,11 +23,9 @@ public class DispatcherServlet extends HttpServlet {
     @Override
     public void init() {
         this.handlerMappings = new HandlerMappings(List.of(
-                new ManualHandlerMapping(),
                 new AnnotationHandlerMapping("com.techcourse.controller")));
         this.handlerAdopters = new HandlerAdopters(List.of(
-                new HandlerExecutionAdopter(),
-                new ControllerAdopter()
+                new HandlerExecutionAdopter()
         ));
     }
